pyproject.toml
[project]
name = "cosmos-mongo-compare"
version = "0.1.0"
description = "Compare sampled documents between Azure Cosmos DB (Mongo API or SQL API) and MongoDB."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pymongo>=4.5",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
cosmos-sql = ["azure-cosmos>=4.5"]
dev = ["pytest>=7.0"]

[project.scripts]
cosmos-mongo-compare = "cosmos_mongo_compare.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

